+++ packages/gov-ui-kit/src/modules/components/proposal/proposalDataListItem/proposalDataListItemStructure/proposalDataListItemStructure.tsx
@@ -5,7 +5,18 @@ import { addressUtils } from '../../../../utils/addressUtils';
 import { ApprovalThresholdResult } from '../approvalThresholdResult';
 import { MajorityVotingResult } from '../majorityVotingResult';
 import { ProposalDataListItemStatus } from '../proposalDataListItemStatus';
-import { type IProposalDataListItemStructureProps } from './proposalDataListItemStructure.api';
+import { type IProposalDataListItemStructureProps, type IPublisher } from './proposalDataListItemStructure.api';
+
+export const maxPublishersDisplayed = 3;
+
+function parsePublisher(publisher: IPublisher, isConnected: boolean, connectedAddress: string | undefined) {
+    const publisherIsConnected = isConnected && addressUtils.isAddressEqual(publisher.address, connectedAddress);
+    const publisherLabel = publisherIsConnected
+        ? 'You'
+        : publisher.name ?? addressUtils.truncateAddress(publisher.address);
+
+    return { label: publisherLabel, link: publisher.link };
+}
 
 /**
  * `ProposalDataListItemStructure` module component
@@ -14,13 +25,13 @@ export const ProposalDataListItemStructure: React.FC<IProposalDataListItemStruct
     const {
         wagmiConfig: config,
         chainId,
+        id,
         className,
         type,
         result,
         date,
-        protocolUpdate,
+        tag,
         publisher,
-        publisherProfileLink,
         status,
         summary,
         title,
@@ -32,35 +43,51 @@ export const ProposalDataListItemStructure: React.FC<IProposalDataListItemStruct
 
     const ongoing = status === 'active' || status === 'challenged' || status === 'vetoed';
 
-    const publisherIsConnected = isConnected && connectedAddress?.toLowerCase() === publisher.address.toLowerCase();
-    const publisherLabel = publisherIsConnected
-        ? 'You'
-        : publisher.name ?? addressUtils.truncateAddress(publisher.address);
+    const parsedPublisher = Array.isArray(publisher)
+        ? publisher.map((p) => parsePublisher(p, isConnected, connectedAddress))
+        : [parsePublisher(publisher, isConnected, connectedAddress)];
+
+    const showParsedPublisher = parsedPublisher.length <= maxPublishersDisplayed;
 
     return (
         <DataList.Item className={classNames('flex flex-col gap-y-4', className)} {...otherProps}>
             <ProposalDataListItemStatus date={date} status={status} voted={voted} />
             <div className="flex flex-col gap-y-1">
-                <p className="line-clamp-1 text-lg leading-tight text-neutral-800 md:text-2xl">{title}</p>
+                <p className="line-clamp-1 flex gap-x-3 text-lg leading-tight md:text-2xl">
+                    {id && <span className="text-neutral-500">{id}</span>}
+                    <span className="text-neutral-800">{title}</span>
+                </p>
                 <p className="line-clamp-2 leading-normal text-neutral-500 md:text-lg">{summary}</p>
             </div>
 
-            {ongoing && type === 'approvalThreshold' && <ApprovalThresholdResult {...result} />}
+            {ongoing && type === 'approvalThreshold' && result && <ApprovalThresholdResult {...result} />}
 
-            {ongoing && type === 'majorityVoting' && <MajorityVotingResult {...result} />}
+            {ongoing && type === 'majorityVoting' && result && <MajorityVotingResult {...result} />}
 
-            <div className="flex items-center gap-x-4 md:gap-x-6">
-                <div className="flex min-h-5 flex-1 items-center gap-x-0.5 text-sm leading-tight text-neutral-600 md:min-h-6 md:text-base">
-                    {/* TODO: apply internationalization [APP-2627] */}
+            <div className="flex items-center justify-between gap-x-4 md:gap-x-6">
+                <div
+                    className={classNames(
+                        'inline-grid auto-cols-auto grid-flow-col content-center',
+                        'min-h-5 gap-x-0.5 text-sm leading-tight text-neutral-600 md:min-h-6 md:gap-x-1 md:text-base',
+                    )}
+                >
                     By
-                    {/* using solution from https://kizu.dev/nested-links/ to nest anchor tags */}
-                    <object type="disregardType" className="ml-0.5 md:ml-1">
-                        <Link href={publisherProfileLink}>{publisherLabel}</Link>
-                    </object>
+                    {showParsedPublisher === false && <span>3+ creators</span>}
+                    {showParsedPublisher &&
+                        parsedPublisher.map(({ label, link }, index) => (
+                            <span key={label} className="truncate">
+                                <object type="disregardType" className="flex shrink">
+                                    {link != null && (
+                                        // using solution from https://kizu.dev/nested-links/ to nest anchor tags
+                                        <Link href={link}>{label}</Link>
+                                    )}
+                                    {link == null && <span className="truncate">{label}</span>}
+                                    {index < parsedPublisher.length - 1 && ','}
+                                </object>
+                            </span>
+                        ))}
                 </div>
-
-                {/* TODO: apply internationalization [APP-2627] */}
-                {protocolUpdate && <Tag label="OSx Update" variant="primary" />}
+                {tag && <Tag label={tag} variant="primary" />}
             </div>
         </DataList.Item>
     );