+++ packages/gov-ui-kit/CHANGELOG.md
@@ -10,10 +10,20 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 ### Added
 
 -   Implement `ProposalDataListItemSkeleton` module component
+-   Extend `addressUtils` with `isAddressEqual` method
 
 ### Changed
 
 -   Remove padding from `DataListContainer`, `DataListFilterStatus`, `DataListPagination` and `DataListRoot`
+-   Add `stageId` and `stageTitle` properties to `IApprovalThresholdResult` & `IMajorityVotingResult` interfaces
+-   Add `id` and optional `tag` properties to `ProposalDataListItemStructure`
+-   Remove `publisherProfileLink` and `protocolUpdate` properties from `ProposalDataListItemStructure`
+-   Update `date` and `result` properties of `ProposalDataListItemStructure` to be optional and `publisher` to allow for
+    multiple publishers
+
+### Fixed
+
+-   `Link` core component to truncate on overflow
 
 ## [1.0.23] - 2024-04-18